feat: add auth to backend

+++ server/src/routes/memories.ts
@@ -3,8 +3,14 @@ import { z } from 'zod'
 import prisma from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (req) => {
+    await req.jwtVerify()
+  })
+  app.get('/memories', async (req) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: req.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -23,7 +29,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (req) => {
+  app.get('/memories/:id', async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -34,9 +40,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     })
+
+    if (memory.userId !== req.user.sub && !memory.isPublic) {
+      return res.status(401).send()
+    }
     return memory
   })
+
   app.post('/memories', async (req) => {
+    const { user } = req
     const bodySchema = z.object({
       coverUrl: z.string(),
       content: z.string(),
@@ -45,7 +57,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const { coverUrl, content, isPublic } = bodySchema.parse(req.body)
     const memory = await prisma.memory.create({
       data: {
-        userId: '0fc9675c-218d-41bc-8fc2-321f8ca9178a',
+        userId: user.sub,
         coverUrl,
         content,
         isPublic,
@@ -54,7 +66,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.put('/memories/:id', async (req) => {
+  app.put('/memories/:id', async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -67,7 +79,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
     const { coverUrl, content, isPublic } = bodySchema.parse(req.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== req.user.sub) {
+      return res.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       data: {
         coverUrl,
         content,
@@ -80,11 +102,21 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (req) => {
+  app.delete('/memories/:id', async (req, res) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
     const { id } = paramsSchema.parse(req.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== req.user.sub) {
+      return res.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
